Removes postcode from Irish Addresses
Postcode (eircode) in Ireland is not appropriate as it is allocated
to an exact address and not to an area or cluster.
This can cause errors when selecting an area on a map and the eircode
for a nearby property is returned.

An example of this is the address 'Forth, Keeloges, County Wexford, Y35
CY80'. The eircode actually points to Knockeen Nursing Home, Knockeen,
Barntown, County Wexford.

For this reason I propose that the eircode/postcode is removed for Irish
addresses